requirements.txt
gymnasium
torch
numpy
pyyaml 
matplotlib
moviepy